Rewrite all data storage in SAK to be object oriented...
And not call eyes of the Elder Gods upon us for what was a
perversion of mathematics and logic.

As a bonus, Shrink mode has been re-written to be key
independent.  Meaning, it no longer requires a specific
kind of selection to work (previously required an equal
number of keys selected per attribute).