SAK: remove compound percentage mode.
I noticed significant bugs in it while doing this re-write (I rolled
back and confirmed that the bugs had always been there).  With this
in mind, I very much doubt this feature was being used by anyone as
these bugs were never reported.  I'm going to nix it, I think the
feature was confusing - creating too many settings to muck with.
I think most of the benefits from this mode can now be gained in
in other ways now that the data storage mechanism in SAK is more
flexible.