Remove non reset-on-apply relative modes in anticipation of new UI
I'm planning on de-cluttering the UI and removing modes I don't
think are used.  This commit does not update the old UI, as its
soon to be replaced.  This affects only the logic code.

There are now only two modes: Absolute and Relative (soon to be renamed)
The fakeReset behavior has been removed.  In its place, if "0" percent
is applied in Relative mode, the curve is reset back to its original
values. This feature was the main reason that compound percentage mode
previous existed at all (and hence why I removed it instead of re-writing
it).