Rewrite slide animation keys UI in QT
A lot of little changes in the process.  I've eliminated a LOT
of settings.  Absolute / Relative modes only exist for quick
pick / hotkey buttons now, but they're integrated MUCH better.
Quick pick buttons regardless of the mode now work seamlessly
with the slider.  I think this is much nicer.  I've eliminated
the ability to set a custom value on the fly - I think now that
the slider works better, it won't be useful.

STILL TODO: Update the Readme; lock 'n hide to QT.